Fix a typo, hints to use autoload, incorrect indentation of module as typed, named blocks where not handled. [ Thanks Drew!]
There was a bug in the auto endcomment for `else and `endif (comment
went to the wrong spot)

Changed some features:
The auto comments used to add comments like:
 endmodule /* module: foo */
 endfunction /* function: bar */

 I deleted the "module:" and "function:" parts as they are rather obvious...

The end for named blocks is now marked.

Added some features:
 Now M-C-t (esc tab) does some reasonable completion;
 it tries to get appropriate keywords based on scope;
 it tries to get tasks based on scope;
 it tires to get modules based on scope.

 eg,
 module foo;
   ini<M-TAB> fills with initial

 module barinsky;
 endmodule

 module foo;

  ba<M-TAB> fills with barinsky

and so on.

However, it still needs more work (vestiges of pascal mode remain).


git-svn-id: file:///verilog-mode-svn/trunk@6 db5bb42b-8904-0410-88e9-aa7026cf2ec3