Teach AUTOs to process escaped backslashes at the end of quoted strings (#1831)
  "...\\..." is a string containing a backslash: \
  "...\"..." is a string containing a double-quote: "

When parsing double-quoted strings, we must detect embedded escaped backslashes
and double-quotes. Detecting when a string terminates with one or more of these
characters is tricky. Strings terminating in "...\"" are not uncommon, and have
reasonably robust support, but strings terminating in "...\\" are quite rare,
and were quite broken.

This patch aims to improve the detection of "...\\" (or "...\\\\", or any even
number of backslashes in a row immediately preceding a double-quote string
terminator.